Made Adding location more user friendly
