+++ app/src/main/java/com/example/cmps121bdd/restroomfinder/MainActivity.java
@@ -1,19 +1,12 @@
 package com.example.cmps121bdd.restroomfinder;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.content.pm.PackageManager;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
-
-import static com.example.cmps121bdd.restroomfinder.R.id.lookupLocation;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
 
@@ -57,7 +50,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 startActivity(intent2);                                                    //starting MapActivity with input
                 break;
             case R.id.addLocation:
-                Intent intent3 = new Intent (this, addLocationDetails.class);    //Creating intent to pass to MapActivity
+                Intent intent3 = new Intent (this, newLocation.class);    //Creating intent to pass to MapActivity
                 startActivity(intent3);                                                    //starting MapActivity with input
                 break;
         }
