+++ app/src/main/java/com/example/cmps121bdd/restroomfinder/addLocationDetails.java
@@ -15,6 +15,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.gms.common.api.ApiException;
@@ -26,50 +27,63 @@ import com.google.android.gms.location.LocationSettingsResponse;
 import com.google.android.gms.location.LocationSettingsStatusCodes;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+
+import java.util.HashMap;
+import java.util.Map;
 
 
 public class addLocationDetails extends AppCompatActivity implements View.OnClickListener{
 
     EditText locName;
+    TextView latitude, longitude;
     Button locOpenMap;
     Button locCoor;
     private boolean mLocationPermissionGranted;
     private static final int LOCATION_PERMISSION_REQUEST_CODE = 1;
     protected static final int REQUEST_CHECK_SETTINGS = 0x1;
 
+    Double lat, lng;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
         setContentView(R.layout.activity_add_location_details);
-
+        Intent bundle = getIntent();
+        String Slat = bundle.getStringExtra("lat");
+        String Slng = bundle.getStringExtra("lng");
+        //Retrive user input coordinates
+        lat = Double.parseDouble(Slat);
+        lng = Double.parseDouble(Slng);
+        //Retrive user input coordinates
         locName = findViewById(R.id.restroom_name);
         locOpenMap = findViewById(R.id.addlocMap);
         locCoor = findViewById(R.id.getLocationBtn);
+        latitude = findViewById(R.id.lat);
+        longitude = findViewById(R.id.lng);
+
+        latitude.setText(Slat);
+        longitude.setText(Slng);
 
         locOpenMap.setOnClickListener(this);
         locCoor.setOnClickListener(this);
-        // Get permission from the User to access their location.
-        getLocationPermission();
 
-        // ASK USER TO ENABLE GPS
-        enableGPS();
-        // ASK USER TO ENABLE GPS
-        displayCoor();
     }
 
     public void onClick(View v) {
         switch(v.getId()) {
 
             case R.id.addlocMap:
-                String location = locName.getText().toString();
-                if (location.equals("")){
+                String inputLocation = locName.getText().toString();
+                if (inputLocation.equals("")){
                     Toast.makeText(this,"Please input a name", Toast.LENGTH_LONG).show();
                     break;
                 }else {
-                    Intent intent1 = new Intent(this, newLocation.class);    //Creating intent to pass to MapActivity
-                    intent1.putExtra("location", location);                    //Adding user input to intent
-                    startActivity(intent1);
+                    FirebaseDatabase database = FirebaseDatabase.getInstance();
+                    final DatabaseReference myRef = database.getReference("Locations");
+                    myRef.child(inputLocation).child("Latitude").setValue(lat);
+                    myRef.child(inputLocation).child("Longitude").setValue(lng);
                     finish();
                     break;
                 }
@@ -93,74 +107,6 @@ public class addLocationDetails extends AppCompatActivity implements View.OnClic
         catch (SecurityException e)
         {
             Toast.makeText(getApplicationContext(), "Can't Add Location Without Permission", Toast.LENGTH_SHORT).show();
-        }
-    }
-
-    //PERMISSIONS STUFF FOR LOCATION----------------------------------------------------------
-    private void getLocationPermission() {
-        /*
-         * Request location permission, so that we can get the location of the
-         * device. The result of the permission request is handled by a callback,
-         * onRequestPermissionsResult.
-         */
-        if (ContextCompat.checkSelfPermission(this.getApplicationContext(),
-                android.Manifest.permission.ACCESS_FINE_LOCATION)
-                == PackageManager.PERMISSION_GRANTED) {
-            mLocationPermissionGranted = true;
-        } else {
-            ActivityCompat.requestPermissions(this,
-                    new String[]{android.Manifest.permission.ACCESS_FINE_LOCATION},
-                    LOCATION_PERMISSION_REQUEST_CODE);
-        }
-    }
-
-    //Check if GPS is enabled, if not, enables
-    public void enableGPS() {
-        LocationRequest locationRequest = LocationRequest.create();
-        locationRequest.setPriority(LocationRequest.PRIORITY_HIGH_ACCURACY);
-        LocationSettingsRequest.Builder builder = new LocationSettingsRequest.Builder()
-                .addLocationRequest(locationRequest);
-        Task<LocationSettingsResponse> result =
-                LocationServices.getSettingsClient(this).checkLocationSettings(builder.build());
-        result.addOnCompleteListener(new OnCompleteListener<LocationSettingsResponse>() {
-            @Override
-            public void onComplete(Task<LocationSettingsResponse> task) {
-                try {
-                    LocationSettingsResponse response = task.getResult(ApiException.class);
-                    // All location settings are satisfied. The client can initialize location
-                    // requests here.
-                } catch (ApiException exception) {
-                    switch (exception.getStatusCode()) {
-                        case LocationSettingsStatusCodes.RESOLUTION_REQUIRED:
-                            // Location settings are not satisfied. But could be fixed by showing the
-                            // user a dialog.
-                            try {
-                                // Cast to a resolvable exception.
-                                ResolvableApiException resolvable = (ResolvableApiException) exception;
-                                // Show the dialog by calling startResolutionForResult(),
-                                // and check the result in onActivityResult().
-                                resolvable.startResolutionForResult(
-                                        addLocationDetails.this,
-                                        REQUEST_CHECK_SETTINGS);
-                            } catch (IntentSender.SendIntentException e) {
-                                // Ignore the error.
-                            } catch (ClassCastException e) {
-                                // Ignore, should be an impossible error.
-                            }
-                            break;
-                        case LocationSettingsStatusCodes.SETTINGS_CHANGE_UNAVAILABLE:
-                            // Location settings are not satisfied. However, we have no way to fix the
-                            // settings so we won't show the dialog.
-                            break;
-                    }
-                }
-            }
-        });
-    }
-    public void checkGPS() {
-        final LocationManager manager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
-        if (!manager.isProviderEnabled(LocationManager.GPS_PROVIDER)) {
-            enableGPS();
         }
     }
