+++ app/src/main/res/layout/activity_add_location_details.xml
@@ -15,13 +15,21 @@
         android:hint="@string/restroom_name"
         tools:layout_editor_absoluteX="109dp"
         tools:layout_editor_absoluteY="16dp" />
+    <TextView
+        android:id="@+id/lat"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
+    <TextView
+        android:id="@+id/lng"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
 
     <Button
         android:id="@+id/addlocMap"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:onClick="onClick"
-        android:text="@string/open_map" />
+        android:text="@string/addLoc" />
 
     <Button
         android:id="@+id/getLocationBtn"
