+++ app/src/main/res/layout/activity_new_location.xml
@@ -1,9 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:map="http://schemas.android.com/apk/res-auto"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/map"
-    android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".newLocation" />
+    android:orientation="vertical"
+    tools:context=".newLocation">
+
+    <fragment
+        android:id="@+id/map"
+        android:name="com.google.android.gms.maps.SupportMapFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".newLocation" />
+
+    <Button
+        android:id="@+id/addlocMap"
+        android:layout_alignBottom="@id/map"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:onClick="addLocMap"
+        android:text="@string/addLoc" />
+
+</RelativeLayout>
